backend/requirements.txt
# framework
fastapi-slim>=0.111.0,<0.112.0

# database
alembic>=1.13.1,<1.14.0
asyncpg==0.29.0
SQLAlchemy>=2.0.30,<2.1.0

# until
orjson>=3.10.3
pydantic-settings>=2.3.0
uvloop>=0.18.0; sys_platform != "win32"

# asgi server
uvicorn>=0.30.0

# tools
pre-commit>=3.5.0
